+++ CfgAmmo.hpp
