+++ CfgCloudlets.hpp
