+++ CfgMagazines.hpp
