--- gl_additions/addons/GLA/scripts/fn_init.sqf
@@ -1,3 +1,0 @@
-player addEventHandler ["fired", {[_this select 6] spawn GLA_fnc_firedHandler;}];
-
-[] spawn GLA_fnc_teargasEffectPlayer;
