--- gl_additions/mod.cpp
@@ -1,12 +1,0 @@
-name = "Grenade Launcher Additions";
-picture = "mod.paa";
-actionName = "";
-action = "";
-description = "GL Additions";
-logo = "mod.paa";
-logoOver = "mod.paa";
-tooltip = "GL Additions";
-tooltipOwned = "";
-overview = "GL Additions";
-author = "HBTurpin";
-
