+++ scripts/fn_dronePlaceCheck.sqf
