+++ scripts/fn_init.sqf
@@ -1,0 +1,3 @@
+player addEventHandler ["fired", {[player, _this select 1, _this select 2, _this select 3, _this select 4, _this select 5, _this select 6, _this select 7] spawn GLA_fnc_firedHandler;}];
+
+[] spawn GLA_fnc_teargasEffectPlayer;
