+++ scripts/fn_teargasEffectAI.sqf
