+++ scripts/fn_teargasEffectPlayer.sqf
